backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client>=1.7.0
asyncpg==0.29.0
psycopg==3.1.15
python-dotenv==1.0.0
openai==1.3.5
google-generativeai==0.4.0
numpy==1.24.3
tiktoken==0.5.1
python-multipart==0.0.6
cryptography==41.0.8
httpx==0.25.2